requirements.txt
ascii-magic
Pillow
colorama
art
pytest
python-dotenv
requests